organized test files a bit
