+++ tests/unit/open_mpic_core/test_mpic_dcv_checker.py
@@ -9,8 +9,8 @@ from open_mpic_core.common_domain.enum.dns_record_type import DnsRecordType
 from open_mpic_core.common_domain.remote_perspective import RemotePerspective
 from open_mpic_core.mpic_dcv_checker.mpic_dcv_checker import MpicDcvChecker
 
-from mock_dns_object_creator import MockDnsObjectCreator
-from valid_check_creator import ValidCheckCreator
+from unit.test_util.mock_dns_object_creator import MockDnsObjectCreator
+from unit.test_util.valid_check_creator import ValidCheckCreator
 
 
 # noinspection PyMethodMayBeStatic
