+++ tests/unit/open_mpic_core/test_mpic_dcv_with_caa_request.py
@@ -4,7 +4,7 @@ import pytest
 from open_mpic_core.common_domain.enum.check_type import CheckType
 from open_mpic_core.mpic_coordinator.domain.mpic_request import MpicDcvWithCaaRequest
 
-from valid_mpic_request_creator import ValidMpicRequestCreator
+from unit.test_util.valid_mpic_request_creator import ValidMpicRequestCreator
 
 
 # noinspection PyMethodMayBeStatic
